move local transport and test suite to core
