+++ packages/core/.eslintrc.cjs
@@ -2,5 +2,6 @@
 module.exports = {
   rules: {
     "@typescript-eslint/no-explicit-any": "off",
+    "@typescript-eslint/no-namespace": "off",
   },
 };
