--- packages/transport-local/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@choreography-ts/transport-local",
-  "version": "1.0.0",
-  "description": "",
-  "main": "dist/index.js",
-  "typings": "dist/index.d.ts",
-  "scripts": {
-    "clean": "rimraf dist",
-    "build": "tsc --project tsconfig.build.json",
-    "test": "jest --silent",
-    "coverage": "jest --silent --coverage",
-    "typecheck": "tsc --noEmit"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@choreography-ts/core": "workspace:*",
-    "@types/jest": "^29.5.1",
-    "jest": "^29.5.0",
-    "ts-jest": "^29.1.0",
-    "typescript": "^5.0.4"
-  },
-  "peerDependencies": {
-    "@choreography-ts/core": "workspace:*"
-  },
-  "dependencies": {
-    "mitt": "^3.0.1"
-  }
-}
