--- packages/transport-local/src/index.ts
@@ -1,2 +1,0 @@
-export { LocalTransport } from "./transport-local.js";
-export type { LocalTransportChannel } from "./transport-local.js";
