--- packages/transport-local/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "dist"
-  },
-  "include": ["src/**/*"],
-  "exclude": ["src/**/*.test.ts"]
-}
