--- packages/transport-test-suite/jest.config.ts
@@ -1,8 +1,0 @@
-import { JestConfigWithTsJest } from "ts-jest";
-import { JestConfigBase } from "../../jest.config.base";
-
-const jestConfig: JestConfigWithTsJest = {
-  ...JestConfigBase,
-};
-
-export default jestConfig;
