--- packages/transport-test-suite/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@choreography-ts/transport-test-suite",
-  "private": true,
-  "version": "1.0.0",
-  "description": "",
-  "main": "dist/index.js",
-  "typings": "dist/index.d.ts",
-  "scripts": {
-    "clean": "rimraf dist",
-    "build": "tsc --project tsconfig.build.json",
-    "test": "jest --silent",
-    "coverage": "jest --silent --coverage",
-    "typecheck": "tsc --noEmit"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/jest": "^29.5.1",
-    "jest": "^29.5.0",
-    "rimraf": "^5.0.1",
-    "ts-jest": "^29.1.0",
-    "typescript": "^5.0.4"
-  },
-  "dependencies": {
-    "@choreography-ts/core": "workspace:*",
-    "@choreography-ts/transport-express": "workspace:*",
-    "@choreography-ts/transport-local": "workspace:*"
-  }
-}
